refactor(toeknVerificationMetadata): add token_verification_metadata requests to dedicated file
